Rename userVault -> vault
