+++ api/resolvers/index.js
@@ -19,7 +19,7 @@ import chainFee from './chainFee'
 import { GraphQLScalarType, Kind } from 'graphql'
 import { createIntScalar } from 'graphql-scalar'
 import paidAction from './paidAction'
-import userVault from './userVault'
+import vault from './vault'
 
 const date = new GraphQLScalarType({
   name: 'Date',
@@ -56,4 +56,4 @@ const limit = createIntScalar({
 
 export default [user, item, message, wallet, lnurl, notifications, invite, sub,
   upload, search, growth, rewards, referrals, price, admin, blockHeight, chainFee,
-  { JSONObject }, { Date: date }, { Limit: limit }, paidAction, userVault]
+  { JSONObject }, { Date: date }, { Limit: limit }, paidAction, vault]
